+++ ichnaea/scripts/map.py
@@ -186,9 +186,13 @@ def generate(db, bucketname, heka_client, stats_client,
             tiles = output
         else:
             tiles = os.path.join(workdir, 'tiles')
-        cmd = ('{enumerate} -z{zoom} {shapes} | xargs -L1 -P{concurrency} '
-               '{render} -o {output} -B 12:0.0379:0.874 -c0088FF -t0 '
-               '-O 16:1600:1.5 -G 0.5{extra}')
+        cmd = ("{enumerate} -z{zoom} {shapes} | xargs -L1 -P{concurrency} "
+               "sh -c 'mkdir -p {output}/$2/$3; {render} "
+               "-B 12:0.0379:0.874 -c0088FF -t0 "
+               "-O 16:1600:1.5 -G 0.5{extra} $1 $2 $3 $4 | "
+               "pngquant --speed=3 --quality=65-95 32 > "
+               "{output}/$2/$3/$4{suffix}.png' dummy")
+
         zoom_0_cmd = cmd.format(
             enumerate=datamaps_enumerate,
             zoom=0,
@@ -196,13 +200,11 @@ def generate(db, bucketname, heka_client, stats_client,
             concurrency=concurrency,
             render=datamaps_render,
             output=tiles,
-            extra=' -T 512')
+            extra=' -T 512',
+            suffix='@2x')
 
         # create high-res version for zoom level 0
         os.system(zoom_0_cmd)
-        zoom_0 = os.path.join(tiles, '0', '0')
-        os.rename(os.path.join(zoom_0, '0.png'),
-                  os.path.join(zoom_0, '0@2x.png'))
 
         zoom_all_cmd = cmd.format(
             enumerate=datamaps_enumerate,
@@ -211,7 +213,8 @@ def generate(db, bucketname, heka_client, stats_client,
             concurrency=concurrency,
             render=datamaps_render,
             output=tiles,
-            extra='')
+            extra='',
+            suffix='')
 
         with stats_client.timer("datamaps.render"):
             os.system(zoom_all_cmd)